Containerize & deploy proxy server
Some other services' endpoints are TBH, hence why some services don't work.
