+++ .babelrc
@@ -1,0 +1,8 @@
+{
+  "presets": [
+    "@babel/preset-react", "@babel/preset-env"
+  ],
+  "plugins": [
+    "babel-plugin-styled-components"
+  ]
+}
