+++ .dockerignore
@@ -1,0 +1,6 @@
+node_modules
+npm-debug.log
+.env
+.git
+test
+.eslintrc.js
