+++ .gitattributes
@@ -1,0 +1,2 @@
+* text=auto
+*.sh text eol=lf
