--- client/src/index.js
@@ -1,5 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import WebsiteBase from './components/WebsiteBase';
-
-ReactDOM.render(<WebsiteBase />, document.getElementById('root'));
