+++ public/index.html
@@ -10,11 +10,11 @@
 <body>
   <div id="root"></div>
   <script src="/bundle.js"></script>
-  <script src="http://localhost:3001/bundle.js"></script>
-  <script src="http://localhost:3002/bundle.js"></script>
+  <script src="http://ec2-3-14-148-108.us-east-2.compute.amazonaws.com:3001/bundle.js"></script>
+  <script src="http://ec2-54-185-79-51.us-west-2.compute.amazonaws.com:3002/bundle.js"></script>
   <script src="http://localhost:3003/main.js"></script>
-  <script src="http://localhost:3004/bundle.js"></script>
-  <script src="http://localhost:3005/bundle.js"></script>
-  <script src="http://localhost:3006/bundle.js"></script>
+  <script src="http://ec2-18-188-192-44.us-east-2.compute.amazonaws.com:3004/bundle.js"></script>
+  <script src="http://ec2-3-14-148-108.us-east-2.compute.amazonaws.com:3005/bundle.js"></script>
+  <script src="http://ec2-54-185-79-51.us-west-2.compute.amazonaws.com:3006/bundle.js"></script>
 </body>
 </html>
